% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare_Data.R
\name{compare_column_names}
\alias{compare_column_names}
\title{Compares the column names of data frames or files in df_list for similarities.
Generates a data frame showing which variable lives where}
\usage{
compare_column_names(df_list, df_names = NULL)
}
\arguments{
\item{df_list}{The list of data frames to compare column names from}

\item{df_names}{The desired names for the data frames. Defaults as NULL}
}
\value{
A data frame that compares all variables across data frames in df_list
}
\description{
Compares the column names of data frames or files in df_list for similarities.
Generates a data frame showing which variable lives where
}
