% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare_Data.R
\name{compare_three}
\alias{compare_three}
\title{Compares three lists for identical values. Can be used to search for identical
observations between data frames.}
\usage{
compare_three(ids1, ids2, ids3, names = c("Dataset1", "Dataset2", "Dataset3"))
}
\arguments{
\item{ids1}{The first column/list to compare}

\item{ids2}{The second column/list to compare}

\item{ids3}{The third column/list to compare}

\item{names}{A list of length 3 of desired names for the data sets}
}
\value{
a list containing:
\itemize{
\item Dataset names
\item num of unique values in each
\item num of common items in each
\item num of values solely in each
\item jaccard similarity
\item overlap percent in both
\item Sample of a common value
\item Sample of unique values in each
}
}
\description{
Compares three lists for identical values. Can be used to search for identical
observations between data frames.
}
