% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_quarto.R
\name{generate_dict_pages}
\alias{generate_dict_pages}
\title{Generates quarto pages from dictionaries in input_dir}
\usage{
generate_dict_pages(
  input_dir,
  output_dir = "dictionary_pages",
  render = FALSE,
  metadata_format = c("csv")
)
}
\arguments{
\item{input_dir}{input path}

\item{output_dir}{output path}

\item{render}{render quarto or not}

\item{metadata_format}{the format of the files being read}
}
\value{
nothing - generates quarto pages
}
\description{
Generates quarto pages from dictionaries in input_dir
}
