% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{generate_meta_data}
\alias{generate_meta_data}
\title{Takes in a data frame and scraped descriptive metadata from the data set.
Used primarily in \code{\link[=save_dict]{save_dict()}}}
\usage{
generate_meta_data(df)
}
\arguments{
\item{df}{The single data frame to gather metadata about}
}
\value{
a list of metadata about a single data frame
}
\description{
Takes in a data frame and scraped descriptive metadata from the data set.
Used primarily in \code{\link[=save_dict]{save_dict()}}
}
\seealso{
\code{\link[=save_dict]{save_dict()}}
}
