% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_quarto.R
\name{generate_page}
\alias{generate_page}
\title{Function for generating a Quarto page for a single metadata-format data
frame. For batch processing, see \code{generate_pages()}}
\usage{
generate_page(
  input_df,
  output_location,
  render = FALSE,
  metadata_format = c("csv")
)
}
\arguments{
\item{input_df}{A named list that contains the single data frame to render.}

\item{output_location}{A path denoting the destination for the rendered file.}

\item{render}{Whether to render the quarto file or not}

\item{metadata_format}{The expected file format of the metadata file}
}
\description{
Function for generating a Quarto page for a single metadata-format data
frame. For batch processing, see \code{generate_pages()}
}
