% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{generate_test_data_dir}
\alias{generate_test_data_dir}
\title{A function that generates a test directory with test files in the desired
directory. Likely to be used in a \code{usethis} test environment with a \code{withr}
tempdir.}
\usage{
generate_test_data_dir(base_dir, test_errors = FALSE)
}
\value{
invisible(nested_path)
}
\description{
A function that generates a test directory with test files in the desired
directory. Likely to be used in a \code{usethis} test environment with a \code{withr}
tempdir.
}
