% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_example}
\alias{get_example}
\title{Gets a non-NA example of the data stored in the column (if applicable)}
\usage{
get_example(df, col_name)
}
\arguments{
\item{df}{The data frame to examine}

\item{col_name}{The column name in the data frame to examine}
}
\value{
a string of the first non-NA value. If all values are NA, returns "All NA"
}
\description{
Gets a non-NA example of the data stored in the column (if applicable)
}
