% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{normalize_name}
\alias{normalize_name}
\title{A function to normalize a name from a path - used in cases where the path of
a file is used in the generation of a unique file name that also shows its
source directory. Primarily used in \code{\link[=save_dict]{save_dict()}}}
\usage{
normalize_name(name)
}
\arguments{
\item{name}{A single string file name that may contain file path conventions
\itemize{
\item desired to be made safe for file naming.
}}
}
\value{
the normalized name
\itemize{
\item Slashes are replaced with an underscore
\item Extensions .csv and .xlsx are removed
}
}
\description{
A function to normalize a name from a path - used in cases where the path of
a file is used in the generation of a unique file name that also shows its
source directory. Primarily used in \code{\link[=save_dict]{save_dict()}}
}
