% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{preview_standardization}
\alias{preview_standardization}
\title{Previews the standardization of variable names, including a comparison between the original and standardized lengths.
To be used in tendem with standardize_item_id}
\usage{
preview_standardization(item_ids, target_length = NULL)
}
\arguments{
\item{item_ids}{A list of values to be standardize}

\item{target_length}{The desired length ot standardization to preview}
}
\value{
Nothing - Instead provides console output with details
}
\description{
Previews the standardization of variable names, including a comparison between the original and standardized lengths.
To be used in tendem with standardize_item_id
}
