% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_metadata.R
\name{save_dict}
\alias{save_dict}
\title{Function to write the data dictionary to files. Can save the
dictionary as a series of CSV files or a single XLSX file with multiple
worksheets. \code{data} is a named list of data frames, most likely returned from
get_files_from_path. This function will always generate the folder
"Dictionary" wherever destination_path is. For example, by default,
\code{destination_path = "."} (the current working directory), so a new folder
"Dictionary" will be added to the working directory.}
\usage{
save_dict(data, export_as = "csv", destination_path = ".")
}
\arguments{
\item{data}{a list of named data frames, likely obtained from
get_files_from_path()}
}
\value{
what it returns
}
\description{
Function to write the data dictionary to files. Can save the
dictionary as a series of CSV files or a single XLSX file with multiple
worksheets. \code{data} is a named list of data frames, most likely returned from
get_files_from_path. This function will always generate the folder
"Dictionary" wherever destination_path is. For example, by default,
\code{destination_path = "."} (the current working directory), so a new folder
"Dictionary" will be added to the working directory.
}
