% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_metadata.R
\name{save_dict_csv}
\alias{save_dict_csv}
\title{Function to save the data dictionary as a directory of CSVs
This function takes in "data" - the list of data frames obtained from
get_files_from_path. Each individual data frame is processed for meta data
and saved as a CSV in a directory that mirrors the directory of its source.}
\usage{
save_dict_csv(data, source_path = "data")
}
\arguments{
\item{data}{The complete data, often as a list of data frames}

\item{source_path}{The path for the data, where the data will be}
}
\value{
Nothing, but saves a workbook as a csv file (.csv).
}
\description{
Function to save the data dictionary as a directory of CSVs
This function takes in "data" - the list of data frames obtained from
get_files_from_path. Each individual data frame is processed for meta data
and saved as a CSV in a directory that mirrors the directory of its source.
}
