% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_metadata.R
\name{save_dict_xlsx}
\alias{save_dict_xlsx}
\title{Save dictionary as XSLX (Deprecated)
Please use \code{\link[=save_dict]{save_dict()}} instead.}
\usage{
save_dict_xlsx(data, source_path = "data")
}
\arguments{
\item{data}{The complete data, often as a list of data frames}

\item{source_path}{The path for the data, where the data will be}
}
\value{
the file path of that the file was saved to
}
\description{
Save dictionary as XSLX (Deprecated)
Please use \code{\link[=save_dict]{save_dict()}} instead.
}
\seealso{
\code{\link[=save_dict]{save_dict()}}
}
