% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{standardize_item_id}
\alias{standardize_item_id}
\title{Standardizes item ids to match in length, adds zeroes mostly}
\usage{
standardize_item_id(item_ids, target_length = NULL)
}
\arguments{
\item{item_ids}{A list of values within a column of a data frame to be standardized for comparison}

\item{target_length}{The desired length of the ID}
}
\value{
Standardized IDs with 0's padding to maintain uniform length
}
\description{
Standardizes item ids to match in length, adds zeroes mostly
}
