% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_metadata.R
\name{save_dict}
\alias{save_dict}
\title{desc}
\usage{
save_dict(data_list)
}
\arguments{
\item{data_list}{a list of named data frames, likely obtained from
get_files_from_path()}
}
\value{
what it returns
}
\description{
desc
}
