% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_metadata.R
\name{save_dict_csv}
\alias{save_dict_csv}
\title{Save dictionary as CSV (Deprecated)
Please use \code{\link[=save_dict]{save_dict()}} instead.}
\usage{
save_dict_csv(data)
}
\arguments{
\item{data}{The complete data, often as a list of data frames}
}
\value{
Nothing, but saves a series of csv files (.csv).
}
\description{
Save dictionary as CSV (Deprecated)
Please use \code{\link[=save_dict]{save_dict()}} instead.
}
\seealso{
\code{\link[=save_dict]{save_dict()}}
}
