% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_metadata.R
\name{save_dict_xlsx}
\alias{save_dict_xlsx}
\title{Function to save the data dictionary as an excel workbook
This function takes in "data" - the list of data frames obtained from
get_files_from_path. Each individual data frame is processed for meta data
and saved as a worksheet to be compiled into a complete workbook}
\usage{
save_dict_xlsx(data, source_path = "data")
}
\arguments{
\item{data}{The complete data, often as a list of data frames}

\item{source_path}{The path for the data, where the data will be}
}
\value{
Nothing, but saves a workbook as an excel file (.xlsx).
}
\description{
Function to save the data dictionary as an excel workbook
This function takes in "data" - the list of data frames obtained from
get_files_from_path. Each individual data frame is processed for meta data
and saved as a worksheet to be compiled into a complete workbook
}
